Allow another photo and text for top-level category item when hovering over (Augment code)
Prompt 1:
I would like for some pictures shown by filters-with-pics to show another image
when user is hovering over the first image with mouse. The original image is coming
from photo-attribute and if the image item info has photo2 attribute
(as already destructured in filters-with-pics), then that should be used as a hover-on image

Prompt 2:
Good! Now lets continue editing filters-with-pics so that if an item has name2 attribute,
then that text is shown on top (centered) on the photo2 image when hovering on the item